UPD: set_external_attributes common method
