+++ mapie/conformity_scores/sets/naive.py
@@ -1,4 +1,4 @@
-from typing import Optional, Tuple, Union
+from typing import Tuple, Union
 
 import numpy as np
 
@@ -9,7 +9,7 @@ from mapie.conformity_scores.sets.utils import (
 from mapie.estimator.classifier import EnsembleClassifier
 
 from mapie._machine_precision import EPSILON
-from mapie._typing import ArrayLike, NDArray
+from mapie._typing import NDArray
 
 
 class Naive(BaseClassificationScore):
@@ -31,30 +31,6 @@ class Naive(BaseClassificationScore):
 
     def __init__(self) -> None:
         super().__init__()
-
-    def set_external_attributes(
-        self,
-        *,
-        classes: Optional[ArrayLike] = None,
-        random_state: Optional[Union[int, np.random.RandomState]] = None,
-        **kwargs
-    ) -> None:
-        """
-        Set attributes that are not provided by the user.
-
-        Parameters
-        ----------
-        classes: Optional[ArrayLike]
-            Names of the classes.
-
-            By default ``None``.
-
-        random_state: Optional[Union[int, RandomState]]
-            Pseudo random number generator state.
-        """
-        super().set_external_attributes(**kwargs)
-        self.classes = classes
-        self.random_state = random_state
 
     def get_conformity_scores(
         self,
